refactor: move sub-components to app folder
