+++ autogpt_platform/frontend/src/app/(platform)/admin/spending/page.tsx
@@ -1,7 +1,7 @@
-import { AdminUserGrantHistory } from "@/components/admin/spending/admin-grant-history-data-table";
 import type { CreditTransactionType } from "@/lib/autogpt-server-api";
 import { withRoleAccess } from "@/lib/withRoleAccess";
 import { Suspense } from "react";
+import { AdminUserGrantHistory } from "./components/AdminUserGrantHistory/AdminUserGrantHistory";
 
 type SpendingDashboardPageSearchParams = {
   page?: string;
