+++ autogpt_platform/frontend/src/app/(platform)/library/agents/[id]/components/AgentRunDetailsView.tsx
@@ -1,9 +1,8 @@
 "use client";
-import React, { useCallback, useMemo } from "react";
 import { isEmpty } from "lodash";
 import moment from "moment";
+import React, { useCallback, useMemo } from "react";
 
-import { useBackendAPI } from "@/lib/autogpt-server-api/context";
 import {
   Graph,
   GraphExecution,
@@ -11,20 +10,25 @@ import {
   GraphExecutionMeta,
   LibraryAgent,
 } from "@/lib/autogpt-server-api";
+import { useBackendAPI } from "@/lib/autogpt-server-api/context";
 
+import ActionButtonGroup from "@/components/agptui/action-button-group";
 import type { ButtonAction } from "@/components/agptui/types";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { IconRefresh, IconSquare } from "@/components/ui/icons";
-import { useToastOnFail } from "@/components/ui/use-toast";
-import ActionButtonGroup from "@/components/agptui/action-button-group";
-import LoadingBox from "@/components/ui/loading";
 import { Input } from "@/components/ui/input";
-
-import {
-  AgentRunStatus,
-  agentRunStatusMap,
-} from "@/components/agents/agent-run-status-chip";
+import LoadingBox from "@/components/ui/loading";
+import { useToastOnFail } from "@/components/ui/use-toast";
 import useCredits from "@/hooks/useCredits";
+import { AgentRunStatus, agentRunStatusMap } from "./AgentRunStatusChip";
+interface Props {
+  agent: LibraryAgent;
+  graph: Graph;
+  run: GraphExecution | GraphExecutionMeta;
+  agentActions: ButtonAction[];
+  onRun: (runID: GraphExecutionID) => void;
+  deleteRun: () => void;
+}
 
 export default function AgentRunDetailsView({
   agent,
@@ -33,14 +37,7 @@ export default function AgentRunDetailsView({
   agentActions,
   onRun,
   deleteRun,
-}: {
-  agent: LibraryAgent;
-  graph: Graph;
-  run: GraphExecution | GraphExecutionMeta;
-  agentActions: ButtonAction[];
-  onRun: (runID: GraphExecutionID) => void;
-  deleteRun: () => void;
-}): React.ReactNode {
+}: Props) {
   const api = useBackendAPI();
   const { formatCredits } = useCredits();
