+++ autogpt_platform/frontend/src/app/(platform)/library/agents/[id]/page.tsx
@@ -1,4 +1,5 @@
 "use client";
+import { useParams, useRouter } from "next/navigation";
 import React, {
   useCallback,
   useEffect,
@@ -6,31 +7,31 @@ import React, {
   useRef,
   useState,
 } from "react";
-import { useParams, useRouter } from "next/navigation";
 
-import { exportAsJSONFile } from "@/lib/utils";
-import { useBackendAPI } from "@/lib/autogpt-server-api/context";
 import {
+  Graph,
   GraphExecution,
   GraphExecutionID,
   GraphExecutionMeta,
-  Graph,
   GraphID,
   LibraryAgent,
   LibraryAgentID,
-  Schedule,
-  ScheduleID,
   LibraryAgentPreset,
   LibraryAgentPresetID,
+  Schedule,
+  ScheduleID,
 } from "@/lib/autogpt-server-api";
+import { useBackendAPI } from "@/lib/autogpt-server-api/context";
+import { exportAsJSONFile } from "@/lib/utils";
 
-import type { ButtonAction } from "@/components/agptui/types";
+import AgentRunDetailsView from "@/app/(platform)/library/agents/[id]/components/AgentRunDetailsView";
+import AgentRunDraftView from "@/app/(platform)/library/agents/[id]/components/AgentRunDraftView";
+import AgentRunsSelectorList from "@/app/(platform)/library/agents/[id]/components/AgentRunsSelectorList";
+import AgentScheduleDetailsView from "@/app/(platform)/library/agents/[id]/components/AgentScheduleDetailsView";
 import DeleteConfirmDialog from "@/components/agptui/delete-confirm-dialog";
-import AgentRunDraftView from "@/components/agents/agent-run-draft-view";
-import AgentRunDetailsView from "@/components/agents/agent-run-details-view";
-import AgentRunsSelectorList from "@/components/agents/agent-runs-selector-list";
-import AgentScheduleDetailsView from "@/components/agents/agent-schedule-details-view";
+import type { ButtonAction } from "@/components/agptui/types";
 import { useOnboarding } from "@/components/onboarding/onboarding-provider";
+import { Button } from "@/components/ui/button";
 import {
   Dialog,
   DialogContent,
@@ -39,9 +40,8 @@ import {
   DialogHeader,
   DialogTitle,
 } from "@/components/ui/dialog";
-import { Button } from "@/components/ui/button";
-import { useToast } from "@/components/ui/use-toast";
 import LoadingBox, { LoadingSpinner } from "@/components/ui/loading";
+import { useToast } from "@/components/ui/use-toast";
 
 export default function AgentRunsPage(): React.ReactElement {
   const { id: agentID }: { id: LibraryAgentID } = useParams();
