+++ autogpt_platform/frontend/src/components/history.ts
@@ -1,6 +1,6 @@
-import { CustomNodeData } from "./CustomNode";
-import { CustomEdgeData } from "./CustomEdge";
 import { Edge } from "@xyflow/react";
+import { CustomEdgeData } from "../app/(platform)/build/components/components/CustomEdge";
+import { CustomNodeData } from "../app/(platform)/build/components/components/CustomNode";
 
 type ActionType =
   | "ADD_NODE"
