+++ autogpt_platform/frontend/src/lib/analytics/GoogleAnalytics.tsx
