+++ autogpt_platform/frontend/src/lib/utils.ts
@@ -1,8 +1,8 @@
 import { type ClassValue, clsx } from "clsx";
 import { twMerge } from "tailwind-merge";
 
+import { NodeDimension } from "@/app/(platform)/build/components/FlowEditor";
 import { Category } from "@/lib/autogpt-server-api/types";
-import { NodeDimension } from "@/components/Flow";
 
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
